Turn gcc warning flags up a notch, and fix some inconsistencies in sign handling

Compiles without warning -- could use testing on hardware, and as much hand-proofreading as we have patience for.

+++ sim/sim_i2c.c
@@ -3,6 +3,7 @@
  *
  */
 #include "marble_api.h"
+#include "sim_api.h"
 #include "i2c_pm.h"
 #include "ltm4673.h"
 #include <stdio.h>
@@ -175,7 +176,7 @@ static int i2c_emu_ltm4673(uint8_t rnw, int reg, uint8_t *data, int len) {
       // LSB-to-MSB
       regval |= (data[n] << 8*(n-offset));
     }
-    printf("Writing 0x%x to page %d, reg 0x%x\r\n", regval, page, reg);
+    printf("Writing 0x%x to page %u, reg 0x%x\r\n", regval, page, (unsigned)reg);
     if (page == 0xff) {
       // Write to all pages
       ltm4673.page0[(uint8_t)(reg & 0xff)] = regval;
